+++ src/inventory.ts
@@ -1,46 +1,4 @@
-
-type potTypes =
-  'cup' |
-  'mug' |
-  'stein' |
-  'bowl' |
-  'mixing bowl' |
-  'plate' |
-  'square plate' |
-  'jug' |
-  'bottle' |
-  'container' |
-  'lidded container' |
-  'raku' |
-  'other'
-
-type colors =
-  'floating blue green' |
-  'cream' |
-  'falls creek shino' |
-  'tenmoku' |
-  'eggplant purple' |
-  'unglazed' |
-  null
-
-type blends =
-  'rim dip' |
-  'diagonal' |
-  'splatter' |
-  null
-
-interface item {
-  id: string,
-  type: potTypes,
-  primaryColor: colors,
-  secondaryColor: colors,
-  blend: blends
-  functional: boolean,
-  photoRoot: string,
-  available: boolean,
-}
-
-interface inventory { inventory: Array<item> }
+import { inventory } from '@/types';
 
 const inv: inventory = {
   inventory: [
